requirements.txt
streamlit
pandas
numpy
matplotlib
altair
python-docx
scipy